# api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Dynamis Code Apps Template API",
    "version": "1.0.0",
    "description": "Stable v1 REST contract. Additive changes preserve authorization, ordering, errors, and side effects. Breaking changes require a new version or published migration window. Query collections use bounded filters, literal search, deterministic sorting, and opaque query-bound cursors. Errors use RFC 9457 Problem Details. See docs/api.md for the human-readable reference."
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "loginLocal",
        "description": "Limited to 10 attempts per source address per minute by default.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}},
        "responses": {
          "200": {"description": "Browser session created", "headers": {"Set-Cookie": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginResponse"}}}},
          "202": {"description": "Multi-factor authentication required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MFALoginResponse"}}}},
          "400": {"$ref": "#/components/responses/Problem"},
          "401": {"$ref": "#/components/responses/Problem"},
          "413": {"$ref": "#/components/responses/Problem"},
          "429": {"$ref": "#/components/responses/RateLimited"}
        }
      }
    },
    "/api/v1/auth/mfa/totp": {
      "post": {
        "operationId": "mfaLoginTOTP",
        "description": "Completes a password login with a time-based authenticator code.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["challenge", "code"], "properties": {"challenge": {"type": "string"}, "code": {"type": "string", "pattern": "^[0-9]{6}$"}}}}}},
        "responses": {"200": {"description": "Browser session created", "headers": {"Set-Cookie": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginResponse"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      },
      "delete": {
        "operationId": "mfaTotpRemove", "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/CSRFToken"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "properties": {"password": {"type": "string", "maxLength": 1024}}}}}},
        "responses": {"204": {"description": "Authenticator factor revoked and affected sessions revoked"}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/mfa/recovery": {
      "post": {
        "operationId": "mfaLoginRecovery",
        "description": "Completes a password login with one single-use recovery code.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["challenge", "code"], "properties": {"challenge": {"type": "string"}, "code": {"type": "string"}}}}}},
        "responses": {"200": {"description": "Browser session created", "headers": {"Set-Cookie": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginResponse"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      }
    },
    "/api/v1/auth/mfa/passkey": {
      "post": {
        "operationId": "mfaLoginPasskey",
        "description": "Completes a password login with a server-verified passkey assertion.",
        "parameters": [{"name": "X-MFA-Challenge", "in": "header", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Browser session created", "headers": {"Set-Cookie": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginResponse"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      }
    },
    "/api/v1/auth/mfa/status": {
      "get": {
        "operationId": "mfaStatus", "security": [{"sessionCookie": []}],
        "responses": {"200": {"description": "MFA status without secrets", "content": {"application/json": {"schema": {"type": "object"}}}}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/mfa/totp/enroll": {
      "post": {
        "operationId": "mfaTotpEnroll", "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/CSRFToken"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "properties": {"password": {"type": "string", "maxLength": 1024}}}}}},
        "responses": {"200": {"description": "One-time TOTP enrollment material", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"type": "object"}}}}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/mfa/totp/complete": {
      "post": {
        "operationId": "mfaTotpComplete", "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/CSRFToken"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["challenge", "code"], "properties": {"challenge": {"type": "string"}, "code": {"type": "string", "pattern": "^[0-9]{6}$"}}}}}},
        "responses": {"200": {"description": "Recovery codes shown once", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"type": "object", "properties": {"recoveryCodes": {"type": "array", "items": {"type": "string"}}}}}}}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/mfa/passkeys/register/options": {
      "post": {
        "operationId": "mfaPasskeyOptions", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/CSRFToken"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Passkey creation options", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"type": "object"}}}}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/mfa/passkeys/register": {
      "post": {
        "operationId": "mfaPasskeyComplete", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/CSRFToken"}, {"name": "X-MFA-Challenge", "in": "header", "required": true, "schema": {"type": "string"}}, {"name": "X-MFA-Name", "in": "header", "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object"}}}},
        "responses": {"200": {"description": "Passkey enrolled; recovery codes shown once", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"type": "object"}}}}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/mfa/passkeys/{passkeyId}": {
      "delete": {
        "operationId": "mfaPasskeyRemove", "security": [{"sessionCookie": []}], "parameters": [{"name": "passkeyId", "in": "path", "required": true, "schema": {"type": "string"}}, {"$ref": "#/components/parameters/CSRFToken"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"password": {"type": "string"}}}}}},
        "responses": {"204": {"description": "Passkey revoked and affected sessions revoked"}, "401": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logoutLocal",
        "security": [{"sessionCookie": []}],
        "parameters": [{"$ref": "#/components/parameters/CSRFToken"}],
        "responses": {
          "204": {"description": "Local session revoked"},
          "400": {"$ref": "#/components/responses/Problem"},
          "401": {"$ref": "#/components/responses/Problem"},
          "403": {"$ref": "#/components/responses/Problem"},
          "429": {"$ref": "#/components/responses/RateLimited"}
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/export": {
      "get": {
        "operationId": "exportWorkspace",
        "description": "Returns a bounded, point-in-time workspace export. Authentication secrets and transient records are excluded.",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {
          "200": {"description": "Workspace export", "headers": {"Content-Disposition": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceExport"}}}},
          "401": {"$ref": "#/components/responses/Problem"},
          "403": {"$ref": "#/components/responses/Problem"},
          "409": {"$ref": "#/components/responses/Problem"},
          "429": {"$ref": "#/components/responses/RateLimited"}
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/files/uploads": {
      "post": {
        "operationId": "initiateFileUpload", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InitiateFileRequest"}}}},
        "responses": {"201": {"description": "Upload initiated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FileResponse"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}, "413": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "session"}
    },
    "parameters": {
...
      "CSRFToken": {"name": "X-CSRF-Token", "in": "header", "required": true, "schema": {"type": "string"}},
...
    },
    "responses": {
      "Problem": {"description": "Problem Details", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
...
      "RateLimited": {"description": "Rate limit exceeded", "headers": {"Retry-After": {"required": true, "schema": {"type": "integer", "minimum": 1}}}, "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
...
      "LoginRequest": {"type": "object", "additionalProperties": false, "required": ["email", "password"], "properties": {"email": {"type": "string", "format": "email", "maxLength": 320}, "password": {"type": "string", "minLength": 1, "maxLength": 1024}}},
      "LoginResponse": {"type": "object", "required": ["csrfToken", "expiresAt"], "properties": {"csrfToken": {"type": "string"}, "expiresAt": {"type": "string", "format": "date-time"}}},
      "MFALoginResponse": {"type": "object", "required": ["mfaRequired", "challenge", "methods"], "properties": {"mfaRequired": {"const": true}, "challenge": {"type": "string", "minLength": 1}, "methods": {"type": "array", "minItems": 1, "items": {"type": "string", "enum": ["passkey", "totp", "recovery"]}}, "passkeyOptions": {"type": ["object", "null"]}}},
...
    }
  }
}